requirements.txt
aiogram>=3.4,<4.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
Telethon>=1.42,<2.0
pytest>=8.0,<9.0
